requirements.txt
requests==2.24.0
Pillow==8.4.0
